# openapi.json
{"openapi":"3.1.0","info":{"title":"Paper-KB API","description":"Paper-KB: fast prototype backend exposing Chroma-powered paper/chunk APIs, search and summarization jobs.","contact":{"name":"Matias Iglesias","email":"you@example.com"},"version":"0.1.0"},"paths":{"/":{"get":{"tags":["health"],"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/_admin/papers_health":{"get":{"tags":["health"],"summary":"Papers Health","description":"Returns a small health summary. Adapter should expose:\n  - backend_name (str)\n  - persisted (bool) or similar\n  - optional counts: n_papers, n_chunks_sample (int)\nAdapter implementations should make best-effort to return this info.","operationId":"papers_health_api__admin_papers_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/papers":{"get":{"tags":["papers"],"summary":"List papers (cache -> storage)","description":"Returns the canonical list of papers. Delegate to services.list_papers(storage).\nThe service will handle caches and normalization.","operationId":"api_list_papers_api_papers_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PapersList"}}}}}}},"/api/papers/{paper_id}":{"get":{"tags":["papers","chunks"],"summary":"Get paginated chunks for a paper (delegates to services.get_paper_chunks)","description":"Request a page of canonical chunks for a paper. Delegates to services.get_paper_chunks(storage, ...).\nThis handler keeps the API contract stable.","operationId":"api_get_paper_api_papers__paper_id__get","parameters":[{"name":"paper_id","in":"path","required":true,"schema":{"type":"string","title":"Paper Id"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Offset"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":200,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaperChunksResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/papers/{paper_id}/chunks/{chunk_id}":{"get":{"tags":["papers","chunks"],"summary":"Get single chunk","operationId":"api_get_chunk_api_papers__paper_id__chunks__chunk_id__get","parameters":[{"name":"paper_id","in":"path","required":true,"schema":{"type":"string","title":"Paper Id"}},{"name":"chunk_id","in":"path","required":true,"schema":{"type":"string","title":"Chunk Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChunkResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/papers/{paper_id}/chunks":{"get":{"tags":["papers","chunks"],"summary":"Get chunks for a paper (optionally filtered by q)","operationId":"api_get_filtered_chunks_api_papers__paper_id__chunks_get","parameters":[{"name":"paper_id","in":"path","required":true,"schema":{"type":"string","title":"Paper Id"}},{"name":"q","in":"query","required":false,"schema":{"type":"string","default":"","title":"Q"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Offset"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":200,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaperChunksResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/_dev/seed":{"post":{"tags":["dev"],"summary":"Dev Seed","operationId":"dev_seed__dev_seed_post","parameters":[{"name":"n_papers","in":"query","required":false,"schema":{"type":"integer","default":8,"title":"N Papers"}},{"name":"min_chunks","in":"query","required":false,"schema":{"type":"integer","default":6,"title":"Min Chunks"}},{"name":"max_chunks","in":"query","required":false,"schema":{"type":"integer","default":12,"title":"Max Chunks"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"ChunkResponse":{"properties":{"chunk_id":{"type":"string","title":"Chunk Id"},"text":{"type":"string","title":"Text"},"chunk_index":{"type":"integer","title":"Chunk Index"},"char_len":{"type":"integer","title":"Char Len"},"header_path":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Header Path"},"pages":{"anyOf":[{"prefixItems":[{"anyOf":[{"type":"integer"},{"type":"null"}]},{"anyOf":[{"type":"integer"},{"type":"null"}]}],"type":"array","maxItems":2,"minItems":2},{"type":"null"}],"title":"Pages"},"meta":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Meta"}},"type":"object","required":["chunk_id","text","chunk_index","char_len"],"title":"ChunkResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"PaperChunksResponse":{"properties":{"paper_id":{"type":"string","title":"Paper Id"},"total":{"type":"integer","title":"Total"},"chunks":{"items":{"$ref":"#/components/schemas/ChunkResponse"},"type":"array","title":"Chunks"}},"type":"object","required":["paper_id","total","chunks"],"title":"PaperChunksResponse"},"PaperMeta":{"properties":{"paper_id":{"type":"string","title":"Paper Id"},"title":{"type":"string","title":"Title"},"authors":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Authors"},"n_chunks":{"type":"integer","title":"N Chunks"},"preview":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Preview"},"pages":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Pages"},"source_file":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Source File"},"created_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Created At"},"pipeline_version":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Pipeline Version"},"embed_model":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Embed Model"}},"type":"object","required":["paper_id","title","n_chunks"],"title":"PaperMeta"},"PapersList":{"properties":{"papers":{"items":{"$ref":"#/components/schemas/PaperMeta"},"type":"array","title":"Papers"}},"type":"object","required":["papers"],"title":"PapersList"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}